requirements.txt
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openpyxl==3.1.5
psycopg2==2.9.9
pycparser==2.22
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.2
